DRY up Capybara feature tests and helper files
