+++ spec/features/user_updates_location_spec.rb
@@ -4,7 +4,7 @@ feature "User updates location" do
   scenario "successfully" do
 
     create_user_valid_credentials
-    add_new_location
+    create_location
     update_location
 
     expect(page).to have_content("Location was successfully updated.")
