--- spec/support/features/update_location.rb
@@ -1,8 +1,0 @@
-module Features
-  def update_location
-    click_on "Update Location"
-    fill_in "Address", with: "101 Main St."
-
-    click_on "Update Location"
-  end
-end
